refactor: componentes de venda
