+++ app/src/main/java/com/anotasmart/ui/screens/carrinho/CarrinhoScreen.kt
@@ -3,10 +3,6 @@ package com.anotasmart.ui.screens.carrinho
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.Remove
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -15,9 +11,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.anotasmart.model.CartItem
 import com.anotasmart.utils.formatSafe
 import com.anotasmart.ui.components.StandardScreen
+import com.anotasmart.ui.screens.carrinho.componentes.CartItemCard
 import com.anotasmart.ui.viewModels.CartViewModel
 
 @Composable
@@ -26,7 +22,7 @@ fun CarrinhoScreen(
     onBackClick: () -> Unit,
     onFinalizarVenda: () -> Unit
 ) {
-    val items by viewModel.items.collectAsState()
+    val items by viewModel.items.collectAsState() // converte flow em stateflow
     val totalValor by viewModel.totalValor.collectAsState()
 
     StandardScreen(
@@ -97,92 +93,3 @@ fun CarrinhoScreen(
     }
 }
 
-@Composable
-fun CartItemCard(
-    item: CartItem,
-    onIncrement: () -> Unit,
-    onDecrement: () -> Unit,
-    onDelete: () -> Unit
-) {
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surfaceVariant
-        ),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = item.nome,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-                Text(
-                    text = "R$ ${formatSafe(item.precoVenda)} x ${item.quantidade} ${item.unidadeMedida.name}",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f)
-                )
-                Text(
-                    text = "Subtotal: R$ ${formatSafe(item.total)}",
-                    style = MaterialTheme.typography.bodySmall,
-                    fontWeight = FontWeight.Bold,
-                    color = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.padding(top = 4.dp)
-                )
-            }
-            
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.End
-            ) {
-                IconButton(onClick = onDecrement) {
-                    Icon(
-                        imageVector = Icons.Default.Remove,
-                        contentDescription = "Remover",
-                        tint = MaterialTheme.colorScheme.primary
-                    )
-                }
-                
-                val displayQuantidade = if (item.unidadeMedida == com.anotasmart.model.UnitType.UN) 
-                                            item.quantidade.toInt().toString() 
-                                        else item.quantidade.toString()
-                                        
-                Text(
-                    text = displayQuantidade,
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier.padding(horizontal = 4.dp)
-                )
-
-                val estoqueDisponivel = item.product?.quantidadeEstoque ?: Double.MAX_VALUE
-                val isService = item.product?.tipoItem == com.anotasmart.model.ItemType.SERVICO
-                val canIncrement = isService || item.quantidade < estoqueDisponivel
-
-                IconButton(
-                    onClick = onIncrement,
-                    enabled = canIncrement
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Add,
-                        contentDescription = "Adicionar",
-                        tint = if (canIncrement) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f)
-                    )
-                }
-                IconButton(onClick = onDelete) {
-                    Icon(
-                        imageVector = Icons.Default.Delete,
-                        contentDescription = "Excluir",
-                        tint = MaterialTheme.colorScheme.error
-                    )
-                }
-            }
-        }
-    }
-}
