+++ app/src/main/java/com/anotasmart/ui/screens/vendas/VendaScreen.kt
@@ -1,10 +1,19 @@
 package com.anotasmart.ui.screens.vendas
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -12,15 +21,15 @@ import androidx.compose.ui.Modifier
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.anotasmart.ui.components.BarraBusca
 import com.anotasmart.ui.components.ListaCategoriasProdutos
-import com.anotasmart.ui.screens.vendas.components.BotaoVendaAvulsa
 import com.anotasmart.ui.screens.vendas.components.DialogAdicionarCarrinho
 import com.anotasmart.ui.screens.vendas.components.DialogItemAvulso
 import com.anotasmart.ui.screens.vendas.components.GradeItems
 import com.anotasmart.ui.viewModels.CartViewModel
 import com.anotasmart.ui.viewModels.CategoriasViewModel
 import com.anotasmart.ui.viewModels.VendaViewModel
-
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
 import com.anotasmart.AnotaSmartApplication
 import com.anotasmart.ui.viewModels.CategoriasViewModelFactory
 import com.anotasmart.ui.viewModels.VendaViewModelFactory
@@ -45,63 +54,77 @@ fun VendaScreen(
     val mostrarDialogItemAvulso by viewModel.mostrarDialogItemAvulso.collectAsState()
     val itensNoCarrinho by cartViewModel.items.collectAsState()
 
-    // Esse box permite a sobreposição feita pelos dialogs adicionar ao carrinho e item avulso
-    Box(modifier = Modifier.fillMaxSize()) {
-        Column(
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(MaterialTheme.colorScheme.background)
+    ) {
+        BarraBusca(
+            query = searchQuery,
+            onQueryChange = viewModel::onSearchQueryChanged
+        )
+
+        // venda avulsa
+        Button(
+            onClick = { viewModel.abrirDialogItemAvulso() },
             modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.background)
+                .fillMaxWidth()
+                .padding(horizontal = 16.dp, vertical = 4.dp)
+                .height(50.dp),
+            shape = RoundedCornerShape(8.dp),
+            colors = ButtonDefaults.buttonColors(
+                containerColor = MaterialTheme.colorScheme.secondaryContainer,
+                contentColor = MaterialTheme.colorScheme.onSecondaryContainer
+            )
         ) {
-            BarraBusca(
-                query = searchQuery,
-                onQueryChange = viewModel::onSearchQueryChanged
+            Icon(
+                imageVector = Icons.Default.Add,
+                contentDescription = "Item avulso",
+                modifier = Modifier.padding(end = 8.dp)
             )
-
-            BotaoVendaAvulsa(
-                onClick = { viewModel.abrirDialogItemAvulso() }
-            )
-
-            ListaCategoriasProdutos(
-                categorias = categorias,
-                categoriaSelecionadaId = categoriaSelecionada,
-                onCategoriaClick = viewModel::onCategoriaSelecionada
-            )
-
-            GradeItems(
-                produtos = produtos,
-                onProdutoClick = { produtoClicado ->
-                    viewModel.selecionarProdutoParaCarrinho(produtoClicado)
-                }
-            )
+            Text(text = "ITEM AVULSO", fontWeight = FontWeight.Bold)
         }
 
-        // Renderização condicional
-        produtoSelecionado?.let { produto ->
-            val quantidadeNoCarrinho = itensNoCarrinho.find { it.product?.id == produto.id }?.quantidade ?: 0.0
-            DialogAdicionarCarrinho(
-                produto = produto,
-                quantidadeNoCarrinho = quantidadeNoCarrinho,
-                onDismissRequest = {
-                    viewModel.limparProdutoSelecionado()
-                },
-                onConfirmar = { quantidade ->
-                    viewModel.adicionarAoCarrinho(produto, quantidade) { item ->
-                        cartViewModel.adicionarItem(item)
-                    }
-                }
-            )
-        }
+        ListaCategoriasProdutos(
+            categorias = categorias,
+            categoriaSelecionadaId = categoriaSelecionada,
+            onCategoriaClick = viewModel::onCategoriaSelecionada
+        )
 
-        // Renderização condicional
-        if (mostrarDialogItemAvulso) {
-            DialogItemAvulso(
-                onDismissRequest = { viewModel.fecharDialogItemAvulso() },
-                onConfirmar = { precoCusto, precoVenda, quantidade ->
-                    viewModel.adicionarItemAvulsoAoCarrinho(precoCusto, precoVenda, quantidade) { item ->
-                        cartViewModel.adicionarItem(item)
-                    }
+        GradeItems(
+            produtos = produtos,
+            onProdutoClick = { produtoClicado ->
+                viewModel.selecionarProdutoParaCarrinho(produtoClicado)
+            }
+        )
+    }
+
+    // Renderização condicional
+    produtoSelecionado?.let { produto ->
+        val quantidadeNoCarrinho = itensNoCarrinho.find { it.product?.id == produto.id }?.quantidade ?: 0.0
+        DialogAdicionarCarrinho(
+            produto = produto,
+            quantidadeNoCarrinho = quantidadeNoCarrinho,
+            onDismissRequest = {
+                viewModel.limparProdutoSelecionado()
+            },
+            onConfirmar = { quantidade ->
+                viewModel.adicionarAoCarrinho(produto, quantidade) { item ->
+                    cartViewModel.adicionarItem(item)
                 }
-            )
-        }
+            }
+        )
+    }
+
+    // Renderização condicional
+    if (mostrarDialogItemAvulso) {
+        DialogItemAvulso(
+            onDismissRequest = { viewModel.fecharDialogItemAvulso() },
+            onConfirmar = { precoCusto, precoVenda, quantidade ->
+                viewModel.adicionarItemAvulsoAoCarrinho(precoCusto, precoVenda, quantidade) { item ->
+                    cartViewModel.adicionarItem(item)
+                }
+            }
+        )
     }
 }
